==> 01_compare_training_and_cohen_kappa.ipynb <==
# TCC: 
# Code: 
# Author: Anísio Pereira Batista Filho

##Essentials
import os
import csv
import numpy as np ##Numpy
import pandas as pd ##Pandas
##Utils
import re
import unicodedata
from datetime import datetime
from itertools import islice
from tqdm.auto import tqdm
import time
import timeit

##Eliminar duplicatas em um dataframe:
def df_drop_duplicates(df):
    df = df.drop_duplicates(subset='tweet_id', keep='first')
    return df

##Comparador de treinamentos (compara dois treinamentos e cria um conjunto de iguais e um de diferentes)
def compare_trainings(dfA, dfB, nome):
    listC = []
    listD = []
    for a in dfA.itertuples():
        for b in dfB.itertuples():
            if a.tweet_id == b.tweet_id:
                if a.label == b.label:
                    dictC = dict({'tweet_id':a.tweet_id,'created_at':a.created_at,'user_location':a.user_location,'tweet_text':a.tweet_text,'label_A':a.label,'label_B':b.label,'pct_certainty_A':a.pct_certainty,'pct_certainty_B':b.pct_certainty})
                    listC.append(dictC)
                else:
                    dictD = dict({'tweet_id':a.tweet_id,'created_at':a.created_at,'user_location':a.user_location,'tweet_text':a.tweet_text,'label_A':a.label,'label_B':b.label,'pct_certainty_A':a.pct_certainty,'pct_certainty_B':b.pct_certainty})
                    listD.append(dictD)     

    dfC = pd.DataFrame(listC, columns=['tweet_id','created_at','user_location','tweet_text','label_A','label_B','pct_certainty_A','pct_certainty_B'])
    dfD = pd.DataFrame(listD, columns=['tweet_id','created_at','user_location','tweet_text','label_A','label_B','pct_certainty_A','pct_certainty_B'])

    dfC.to_csv(f"data/corpus_labeled/iguais/{nome}[iguais].csv", sep=",", index=False)
    dfD.to_csv(f"data/corpus_labeled/diferentes/{nome}[diferentes].csv", sep=",", index=False)

##Calculo de Cohen Kappa:
def cohen_kappa(df, dfA, dfB):
    ##Criando dicionário de rótuloxquantidade com as rotulações do avaliador A:
    labelerA = dict()
    fake_news = opinion = news = 0
    for row in dfA.itertuples():
        if row.label == 1:
            fake_news += 1
            labelerA['fake_news'] = fake_news
        else: 
            if row.label == 0:
                opinion += 1
                labelerA['opinion'] = opinion
            else:
                if row.label == -1:
                    news += 1
                    labelerA['news'] = news
    print('Rotulador A: ',labelerA)

    ##Calculando a probabilidade de concordancia randomica do avaliador A:
    peA = dict()
    peA['fake_news'] = labelerA['fake_news']/dfA.shape[0]
    peA['opinion'] = labelerA['opinion']/dfA.shape[0]
    peA['news'] = labelerA['news']/dfA.shape[0]
    print('peA: ',peA)

    ##Criando dicionário de rótuloxquantidade com as rotulações do avaliador B:
    labelerB = dict()
    fake_news = opinion = news = 0
    for row in dfB.itertuples():
        if row.label == 1:
            fake_news += 1
            labelerB['fake_news'] = fake_news
        else: 
            if row.label == 0:
                opinion += 1
                labelerB['opinion'] = opinion
            else:
                if row.label == -1:
                    news += 1
                    labelerB['news'] = news
    print('Rotulador B: ',labelerB)

    ##Calculando a probabilidade de concordancia randomica do avaliador B:
    peB = dict()
    peB['fake_news'] = labelerB['fake_news']/dfB.shape[0]
    peB['opinion'] = labelerB['opinion']/dfB.shape[0]
    peB['news'] = labelerB['news']/dfB.shape[0]
    print('peB: ',peB)

    ##Calculando a probabilidade de ambos os avaliadores (para fake_news, opinion e news):
    pe = dict()
    pe['fake_news'] = peA['fake_news']*peB['fake_news']
    pe['opinion'] = peA['opinion']*peB['opinion']
    pe['news'] = peA['news']*peB['news']
    print('pe:',pe)

    ##Calculando a probabilidade de aceitação total:
    Pr = pe['fake_news']+pe['opinion']+pe['news']
    print('Pr: ',Pr)

    ##Calculando Po
    Po = df.shape[0]/dfA.shape[0]
    print('Po: ',Po)
    ##Calculando Kappa:
    K = (Po - Pr)/(1 - Pr)
    print('K: ',K)

## Função main:

### Rotulação 1

###Abertura de arquivo e criação do dataframe:
df1_file = pd.read_csv("data/original_corpus_to_label/base_para_rotulaçao1.csv", sep=",")
df1 = pd.DataFrame(df1_file)
df1 = df1.drop_duplicates(subset='tweet_id', keep='first')
print("Tamanho:", df1.shape[0])
df1.sample(5)

###Abertura de arquivo e criação do dataframe:
dfA1_file = pd.read_csv("data/corpus_labeled/rotulaçao1[anisiofilho].csv", sep=",")
dfA1 = pd.DataFrame(dfA1_file)
dfA1 = dfA1.drop_duplicates(subset='tweet_id', keep='first')
print("Tamanho:", dfA1.shape[0])
dfA1.sample(5)

###Verificação da quantidade de tuítes por rótulo
dfA1.groupby('label').size()

###Abertura de arquivo e criação do dataframe:
dfB1_file = pd.read_csv("data/corpus_labeled/rotulaçao1[debora].csv", sep=",")
dfB1 = pd.DataFrame(dfB1_file)
dfB1 = dfB1.drop_duplicates(subset='tweet_id', keep='first')
print("Tamanho:", dfB1.shape[0])
dfB1.sample(5)

###Verificação da quantidade de tuítes por rótulo
dfB1.groupby('label').size()

### Chamadas de funções:

compare_trainings(dfA1,dfB1,'rotulaçao1')

###Abertura de arquivo e criação do dataframe:
dfC1_file = pd.read_csv("data/corpus_labeled/iguais/rotulaçao1[iguais].csv", sep=",")
dfC1 = pd.DataFrame(dfC1_file)
dfC1 = dfC1.drop_duplicates(subset='tweet_id', keep='first')
print("Tamanho:", dfC1.shape[0])
dfC1.sample(5)

dfC1.groupby('label_A').size()
#dfC1.groupby('label_B').size()

###Abertura de arquivo e criação do dataframe:
dfD1_file = pd.read_csv("data/corpus_labeled/diferentes/rotulaçao1[diferentes].csv", sep=",")
dfD1 = pd.DataFrame(dfD1_file)
dfD1 = dfD1.drop_duplicates(subset='tweet_id', keep='first')
print("Tamanho:", dfD1.shape[0])
dfD1.sample(5)

dfD1.groupby('label_A').size()

dfD1.groupby('label_B').size()

cohen_kappa(dfC1, dfA1, dfB1)

### Rotulação 2

###Abertura de arquivo e criação do dataframe:
df2_file = pd.read_csv("data/original_corpus_to_label/base_para_rotulaçao2.csv", sep=",")
df2 = pd.DataFrame(df2_file)
df2 = df.drop_duplicates(subset='tweet_id', keep='first')
print("Tamanho:", df2.shape[0])
df2.sample(5)

###Abertura de arquivo e criação do dataframe:
dfA2_file = pd.read_csv("data/corpus_labeled/rotulaçao2[anisiofilho].csv", sep=",")
dfA2 = pd.DataFrame(dfA2_file)
dfA2 = dfA2.drop_duplicates(subset='tweet_id', keep='first')
print("Tamanho:", dfA2.shape[0])
dfA2.sample(5)

dfA2.groupby('label').size()

###Abertura de arquivo e criação do dataframe:
dfB2_file = pd.read_csv("data/corpus_labeled/rotulaçao2[andre].csv", sep=",")
dfB2 = pd.DataFrame(dfB2_file)
dfB2 = dfB2.drop_duplicates(subset='tweet_id', keep='first')
print("Tamanho:", dfB2.shape[0])
dfB2.sample(5)

dfB2.groupby('label').size()

### Chamadas de funções:

compare_trainings(dfA2,dfB2,'rotulaçao2')

dfC2_file = pd.read_csv("data/corpus_labeled/iguais/rotulaçao2[iguais].csv", sep=",")
dfC2 = pd.DataFrame(dfC2_file)
dfC2 = dfC2.drop_duplicates(subset='tweet_id', keep='first')
print("Tamanho:", dfC2.shape[0])
dfC2.sample(5)

dfC2.groupby('label_A').size()
#dfC.groupby('label_B').size()

dfD2_file = pd.read_csv("data/corpus_labeled/diferentes/rotulaçao2[diferentes].csv", sep=",")
dfD2 = pd.DataFrame(dfD2_file)
dfD2 = dfD2.drop_duplicates(subset='tweet_id', keep='first')
print("Tamanho:", dfD2.shape[0])
dfD2.sample(5)

dfD2.groupby('label_A').size()

dfD2.groupby('label_B').size()

cohen_kappa(dfC2, dfA2, dfB2)

## Unindo dataframes

###Rotulação iguais
frames_iguais = [dfC1, dfC2]

result_iguais = pd.concat(frames_iguais)
print("Tamanho:", result_iguais.shape[0])
result_iguais.sample(5)

result_iguais.to_csv(f"data/corpus_labeled/iguais/rotulacao_uniao[iguais].csv", sep=",", index=False)

###Rotulação diferentes
frames_diferentes = [dfD1, dfD2]

result_diferentes = pd.concat(frames_diferentes)
print("Tamanho:", result_diferentes.shape[0])
result_diferentes.sample(5)

result_diferentes.to_csv(f"data/corpus_labeled/diferentes/rotulacao_uniao[diferentes].csv", sep=",", index=False)

df_completo_iguais = pd.read_csv("data/corpus_labeled/iguais/rotulacao_uniao[iguais].csv", sep=",")

df_completo_iguais.groupby('label_A').size()

df_completo_iguais.groupby('label_B').size()

### Miscelanea

#
if dfA1.tweet_id is not dfB1.tweet_id:
    print(dfA1.tweet_id)

dfA1 = dfA1.assign(InDf2=dfA1.tweet_id.isin(dfB1.tweet_id).astype(int))

dfA1.assign(InDf2=dfA1.tweet_id.isin(dfB1.tweet_id).astype(int))

dfA1.groupby('InDf2').size()

dfA1.loc[dfA1['InDf2'] == 0]
