==> 03_geracao_caracteristicas.ipynb <==
# TCC: 
# Code: 
# Author: Anísio Pereira Batista Filho

## Importações

import csv
import matplotlib
import matplotlib.pyplot as plt
import numpy as np
import os
import pandas as pd
import re
import spacy
import string
import time
import tweepy as tw
import unicodedata

from datetime import datetime
from geopy.extra.rate_limiter import RateLimiter
from geopy.geocoders import Nominatim
from itertools import islice
import plotly.express as px
from spellchecker import SpellChecker
from tqdm.auto import tqdm
from unidecode import unidecode
from wordcloud import WordCloud, ImageColorGenerator

## Abertura de arquivo e criação do dataframe

df = pd.read_csv('data/corpus_labeled/iguais/bases_tcc/02_text_preprocessing_base.csv', sep=",", low_memory=False)

df_estados = pd.read_csv("data/utils/abreviações_estados.csv", sep=",", low_memory=True)

dict_estados = df_estados.set_index('sigla')['estado'].to_dict()

df_estados_regioes = pd.read_csv("data/utils/estados_abreviações_regiões.csv", sep=",", low_memory=True)

dict_estados_regioes =  df_estados_regioes.set_index('estado')['região'].to_dict()

## Funções

### Inicializações

tqdm.pandas()

# Twitter Developer Keys for crawler_pivic app
consumer_key = '<consumer_key>'
consumer_secret = '<consumer_secret>'
access_token = '<access_token>'
access_token_secret = '<access_token_secret>'

### Autgenticação entre Twitter Developer e este script:
auth = tw.OAuthHandler(consumer_key, consumer_secret)
auth.set_access_token(access_token, access_token_secret)
api = tw.API(auth, wait_on_rate_limit=True)#, wait_on_rate_limit_notify=True)

spell = SpellChecker(language='pt')

geolocator = Nominatim(user_agent="google")
geocode = RateLimiter(geolocator.geocode, min_delay_seconds=1)

nlp = spacy.load('pt_core_news_lg')

### Selecionador de turno

def seleciona_turno(linha):
    linha = pd.to_datetime(linha)
    linha = linha.strftime("%H:%M:%S")
    if linha > '00:00:00' and linha < '11:59:59':
        return 'manhã'
    elif linha > '12:00:00' and linha < '17:59:59':
        return 'tarde'
    else: 
        return 'noite'

### Crawler de usuários

def user_crawler(linha):
    try:
        tweetFetched = api.get_status(linha)
        #Para saber o que poder ser recuperado pelo objeto status ou objeto user, pesquisar.       
        user_screen_name = tweetFetched.user.screen_name
        user_id = tweetFetched.user.id_str
        time.sleep(2)        
    except:
        user_screen_name = 'invalid_user'
        user_id = 'invalid_user'
        
    return user_screen_name, user_id

### Tratamento da coluna de localização

def trata_localizacao(linha):
    ### Convertendo para string
    linha = str(linha)
    ### Removendo os '-'
    linha = re.sub(r'[^\w\s]', ' ', linha)
    ### Removendo a pontuação da string
    linha = linha.translate(str.maketrans('', '', string.punctuation))
    ### Colocando as palavras em caixa baixa
    linha = linha.lower()
    ### Aplicando a correção gramatical do Spellchecker
    linha = spell.correction(linha)
    ### Removendo a acentuação da string
    linha = unidecode(linha)
    ### Removendo espaços múltiplos
    linha = re.sub(r'\s+', ' ', linha)

    for parte in linha.split():
        if parte in dict_estados.keys():
            linha = linha.replace(parte, dict_estados[parte])
    return linha

### Gerador de listas de estados e regiçoes utilizando Geopy

def gera_estado_regiao_geopy(linha):
    try:
        local = geolocator.geocode(linha, addressdetails=True)
        local = local.raw
        local = local['address']

        if local['country_code'] == 'br':
            try:
                if local['state'] in dict_estados_regioes:
                    estado = local['state']
                    regiao = dict_estados_regioes[local['state']]
            except:
                estado = 'statenotdefined'
                regiao = 'regionnotdefined'
        else:
            estado = 'notbrazilstate'
            regiao = 'notbrazilregion'
    except:
        #local = {'state': 'invalid', 'region': 'invalid', 'country': 'invalid'}
        estado = 'invalidstate'
        regiao = 'invalidregion'

    return estado, regiao

### Contador de caracteres

def conta_caracteres(linha):    
    return len(linha)

### Contador de palavras

def conta_palavras(linha):
    return len(linha.split())

### Analisador morfológico

def analisa_morfologia(linha):
    doc = nlp(linha)
    
    substantivos = adjetivos = verbos = adverbios = 0
    for token in doc:
        if token.pos_ == 'NOUN':
            substantivos += 1
        elif token.pos_ == 'ADJ':
            adjetivos += 1
        elif token.pos_ == 'VERB':
            verbos += 1
        elif token.pos_ == 'ADV':
            adverbios += 1
    
    return substantivos, adjetivos, verbos, adverbios

## Main

### Criação de colunas e ajuste de dados

#### Criação da coluna de turnos

df['time_shift'] = df.created_at.progress_map(seleciona_turno)

#### Criação das colunas de nome de usuário e id de usuário

df['user_screen_name'], df['user_id'] = zip(*df.tweet_id.progress_apply(lambda x: user_crawler(x)))

#### Preenchimento dos campos em branco da localização dos usuários

df['user_location'] = df['user_location'].fillna("invalidlocation")

#### Criação da coluna de localização tratada

df['location_treated'] = df.user_location.progress_map(trata_localizacao)

#### Criação das colunas de Estado e Região com base na localização tratada

df['state_location'], df['region_location'] = zip(*df.location_treated.progress_apply(lambda x: gera_estado_regiao_geopy(x)))

#### Calculo da quantidade de caracteres do stemming

df['tweet_text_stemming_char_len'] = df.tweet_text_stemming.progress_map(conta_caracteres)

#### Calculo da quantidade de palavras do stemming

df['tweet_text_stemming_word_len'] = df.tweet_text_stemming.progress_map(conta_palavras)

#### Calculo da quantidade de caracteres do lemmatization

df['tweet_text_lemmatization_char_len'] = df.tweet_text_lemmatization.progress_map(conta_caracteres)

#### Calculo da quantidade de palavras do lemmatization

df['tweet_text_lemmatization_word_len'] = df.tweet_text_lemmatization.progress_map(conta_palavras)

#### Geração de características NAVA para o stemming

df['tweet_text_stemming_noun'], df['tweet_text_stemming_adj'], df['tweet_text_stemming_verb'], df['tweet_text_stemming_adv']  = zip(*df.tweet_text_stemming.progress_apply(lambda x: analisa_morfologia(x)))

#### Geração de características NAVA para o lemmatization

df['tweet_text_lemmatization_noun'], df['tweet_text_lemmatization_adj'], df['tweet_text_lemmatization_verb'], df['tweet_text_lemmatization_adv']  = zip(*df.tweet_text_lemmatization.progress_apply(lambda x: analisa_morfologia(x)))

## Salvando a base

df.to_csv('data/corpus_labeled/iguais/bases_tcc/03_geracao_carcteristicas_base.csv', sep=',', index=False)

df
